[SPARK-52729][SQL][FOLLOWUP] Introduce RelationInfo as a common base for TableInfo and ViewInfo
Make TableInfo and ViewInfo sibling subclasses of a new common base RelationInfo
instead of having ViewInfo extend TableInfo, so views no longer inherit the
table-only surface (partitioning, constraints, provider, location).
